pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "cafe oms"
authors = [
    {name = "i.sabirov"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
